Don't copy/splice the todos array on edit
Currently, editing a todo (toggling it or changing its text) creates a new Todo object and splices it into the array. This is unnecessary work, if we instead make the `<todo-item>` take each field of the todo separately (common in Lit for leaf elements) then we can update it in place. This exercises more idiomatic Lit code (more template bindings) and less copying and splicing of the array of todos.