Keep the parameter 'tags' as an array in the body of the function
Previously, we were transforming it into a set. This was giving problems
in a later condition, so that no message/debugInfo were defined, and the
user would get an alert with the message "undefined".

Now the set stays local to the code block where it's used, so that the
parameter variable "tags" is an array for the whole function, for more
consistency with the "names" variable.