Add debug features
- Subdomain running options:
  - hosted: using `-$ITER` subdomain suffix
  - local: using increasing port numbers
- Add measure prepare phase for "loading" style metric
- Move the warmup suite to the main config
- Display all top level metrics directly in the details view


The subdomain runner works by the outer default runner communicating
with the inner (embedded) runner via postMessage. There are 2 new
runners

- SubdomainBenchmarkRunner: the outer runner
- EmbeddedBenchmarkRunner: the inner runner

The main page switches over to the SubdomainBenchmarkRunner if the
corresponding param is set.
The EmbeddedBenchmarkRunner is then used by setting `?embedded` and
we end up with double wrapped workload (easy right...):

  SubdomainBenchmarkRunner:
    iframe:
      EmbeddedBenchmarkRunner:
        iframe:
           Suite

This approach doesn't require changing the current workloads and we
can still easily measure framework setup time. Hopefully this can
be cleaned up in the future by refactoring the runner.

Change-Id: Ib813e8d83fd164fb2849b6889f187c1904b21965