Add TodoMVC written in Dart using Jaspr DOM framework
This adds a TodoMVC app written in Dart using the Jaspr DOM framework.
The app is compiled in two variants: dart2js and dart2wasm.

We compile with

  * dart2js in unsound (-O4) mode
  * dart2wasm in sound (-O2) mode

as those are the default settings in e.g. flutter tooling.

To remember this difference we specify the optimization mode in
the benchmark name.

The app can be compiled by installing the Dart SDK and running

```
experimental/todomvc-dart-jaspr % ./build.sh |& tee build.log
...
```